feat: harden library for v1.0.0
Types & correctness:
- pass an initial value to every useRef (fixes 4 React 19 type errors)
- give useIntersectionObserver a real lazy conditional return type, a stable
  callback ref, and number[] threshold support
- fix useLazyState previousValue (was always undefined)
- cancel pending timers on unmount in useDebouncedCallback and useResizeObserver

Packaging:
- collapse packages/react into src/ (removes the duplicate package name)
- dual ESM + CJS build via tsdown unbundle with per-file "use client" so
  useObjectFit stays usable in Server Components
- inline the resize emitter and drop nanoevents -> zero runtime dependencies
- drop the unused react-dom peer; set peer react >=18; add engines.node >=18

Tooling & tests:
- Biome-only (remove the dead ESLint config); add typecheck/lint/test scripts
- add a bun:test suite: render smoke, SSR safety, debounce leak regression
- rewrite publish CI for bun; add a CI workflow with a React 18/19 matrix
