--- .eslintrc.json
@@ -1,18 +1,0 @@
-{
-  "extends": ["eslint:recommended", "plugin:react/recommended", "plugin:react-hooks/recommended"],
-  "plugins": ["react", "react-hooks"],
-  "parserOptions": {
-    "ecmaVersion": 2020,
-    "sourceType": "module",
-    "ecmaFeatures": {
-      "jsx": true
-    }
-  },
-  "env": {
-    "browser": true,
-    "es2015": true
-  },
-  "rules": {
-    "react/react-in-jsx-scope": "off"
-  }
-}
