+++ package.json
@@ -1,24 +1,24 @@
 {
   "name": "hamo",
-  "version": "1.0.0-dev.10",
-  "description": "hamo means hook, do the math.",
+  "version": "1.0.0",
+  "description": "A collection of performance-oriented React hooks.",
   "type": "module",
-  "workspaces": [
-    "packages/*",
-    "playground",
-    "playground/*"
-  ],
   "sideEffects": false,
-  "unpkg": "./dist/hamo.mjs",
-  "main": "./dist/hamo.mjs",
+  "main": "./dist/hamo.cjs",
   "module": "./dist/hamo.mjs",
   "types": "./dist/hamo.d.ts",
   "exports": {
     ".": {
-      "types": "./dist/hamo.d.ts",
-      "default": "./dist/hamo.mjs"
+      "import": {
+        "types": "./dist/hamo.d.ts",
+        "default": "./dist/hamo.mjs"
+      },
+      "require": {
+        "types": "./dist/hamo.d.cts",
+        "default": "./dist/hamo.cjs"
+      }
     },
-    "./dist/*": "./dist/*"
+    "./package.json": "./package.json"
   },
   "files": [
     "dist"
@@ -31,39 +31,55 @@
   "homepage": "https://github.com/darkroomengineering/hamo",
   "repository": {
     "type": "git",
-    "url": "https://github.com/darkroomengineering/hamo"
+    "url": "git+https://github.com/darkroomengineering/hamo.git"
+  },
+  "engines": {
+    "node": ">=18"
   },
   "funding": {
     "type": "github",
     "url": "https://github.com/sponsors/darkroomengineering"
   },
+  "workspaces": [
+    "playground"
+  ],
   "scripts": {
     "build": "tsdown",
     "dev": "bun run --parallel dev:build dev:playground",
     "dev:build": "tsdown --watch",
     "dev:playground": "bun --filter playground dev",
-    "version:dev": "npm version prerelease --preid dev --force --no-git-tag-version",
+    "typecheck": "tsc --noEmit",
+    "lint": "biome check",
+    "format": "biome format --write",
+    "test": "bun test",
+    "prepublishOnly": "tsdown",
     "version:patch": "npm version patch --force --no-git-tag-version",
     "version:minor": "npm version minor --force --no-git-tag-version",
     "version:major": "npm version major --force --no-git-tag-version",
     "postversion": "bun run build",
-    "publish:main": "npm publish",
-    "publish:dev": "npm publish --tag dev"
+    "publish:main": "npm publish"
   },
   "keywords": [
     "react",
-    "hooks"
+    "hooks",
+    "react-hooks",
+    "resize-observer",
+    "intersection-observer",
+    "media-query",
+    "ssr"
   ],
   "devDependencies": {
-    "@biomejs/biome": "1.9.4",
+    "@biomejs/biome": "^2.2.0",
+    "@happy-dom/global-registrator": "^15.0.0",
+    "@testing-library/react": "^16.1.0",
+    "@types/react": "^19.0.0",
+    "happy-dom": "^15.0.0",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
     "tsdown": "^0.21.4",
     "typescript": "^5.4.5"
   },
-  "dependencies": {
-    "nanoevents": "^9.0.0"
-  },
   "peerDependencies": {
-    "react": ">=17.0.0",
-    "react-dom": ">=17.0.0"
+    "react": ">=18.0.0"
   }
 }
