--- packages/react/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "hamo",
-  "type": "module",
-  "devDependencies": {
-    "react": "^18.2.0",
-    "@types/react": "^18.2.77"
-  }
-}
