--- packages/react/src/use-rect/emitter.ts
@@ -1,3 +1,0 @@
-import { createNanoEvents } from 'nanoevents'
-
-export const emitter = createNanoEvents()
