+++ src/use-intersection-observer/index.ts
@@ -1,39 +1,49 @@
+'use client'
+
 import { useCallback, useEffect, useRef, useState } from 'react'
 
 /**
  * @name useIntersectionObserver
- * @description A React hook that oberves element visibility using IntersectionObserver.
- * @param {HTMLElement} root (optional)
+ * @description A React hook that observes element visibility using IntersectionObserver.
+ * @param {Element | Document} root (optional)
  * @param {string} rootMargin (optional, default: `0px`)
- * @param {number} threshold (optional, default: `0`)
+ * @param {number | number[]} threshold (optional, default: `0`)
  * @param {boolean} once (optional, default: `false`)
  * @param {boolean} lazy (optional, default: `false`)
  * @param {function} callback (optional)
  * @param {array} deps (optional)
- * @returns {array} [setElement, entry]
+ * @returns {array} [setElement, lazy ? getEntry : entry]
  */
 
-export function useIntersectionObserver(
+export function useIntersectionObserver<L extends boolean = false>(
   {
     root = null,
     rootMargin = '0px',
     threshold = 0,
     once = false,
-    lazy = false,
+    lazy = false as L,
     callback = () => {},
   }: {
-    root?: HTMLElement | null
+    root?: Element | Document | null
     rootMargin?: string
-    threshold?: number
+    threshold?: number | number[]
     once?: boolean
-    lazy?: boolean
+    lazy?: L
     callback?: (entry: IntersectionObserverEntry | undefined) => void
   } = {},
-  deps = []
-) {
-  const entryRef = useRef<IntersectionObserverEntry>()
+  deps: any[] = []
+): [
+  (element: HTMLElement | null) => void,
+  L extends true
+    ? () => IntersectionObserverEntry | undefined
+    : IntersectionObserverEntry | undefined,
+] {
+  const entryRef = useRef<IntersectionObserverEntry | undefined>(undefined)
   const [entry, setEntry] = useState<IntersectionObserverEntry>()
-  const [element, setElement] = useState<HTMLElement | null>()
+  const [element, setElement] = useState<HTMLElement | null>(null)
+
+  const callbackRef = useRef(callback)
+  callbackRef.current = callback
 
   useEffect(() => {
     if (!element) return
@@ -46,7 +56,7 @@ export function useIntersectionObserver(
           setEntry(entry)
         }
 
-        callback(entry)
+        callbackRef.current(entry)
 
         if (once && entry?.isIntersecting) intersection.disconnect()
       },
@@ -63,10 +73,12 @@ export function useIntersectionObserver(
     }
   }, [element, root, rootMargin, threshold, lazy, once, ...deps])
 
-  const get = useCallback(() => entryRef.current, [])
+  const getEntry = useCallback(() => entryRef.current, [])
 
-  return [
-    setElement,
-    (lazy ? get : entry) as IntersectionObserverEntry, // fix type
-  ] as const
+  return [setElement, lazy ? getEntry : entry] as [
+    (element: HTMLElement | null) => void,
+    L extends true
+      ? () => IntersectionObserverEntry | undefined
+      : IntersectionObserverEntry | undefined,
+  ]
 }
