+++ src/use-rect/utils.ts
